+++ src/frontend/web_application/src/scenes/ContactBook/components/ImportContactForm/style.scss
@@ -1,0 +1,19 @@
+@import '../../../../styles/common';
+@import '../../../../styles/object/o-clickable';
+@import '../../../../styles/object/o-form-element';
+@import '../../../../styles/object/o-hidden-element';
+@import '../../../../styles/vendor/bootstrap_foundation-sites';
+
+.m-import-contact-form {
+  width: map-get($co-form__width, 'xlarge');
+
+  &__buttons {
+    @include flex-grid-row;
+    justify-content: space-between;
+    margin-top: $co-margin;
+  }
+
+  &__button {
+    @include flex-grid-column(shrink);
+  }
+}
